Extract getFieldOptions
